pyproject.toml
[tool.poetry]
name = "repl_python3_Tohru"
version = "0.1.0"
description = ""
authors = ["bbrez <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"
"discord-py-slash-command" = "^1.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
